efaas-automation/requirements.txt
cachetools==3.1.0
google-auth==1.6.3
pyasn1==0.4.5
pyasn1-modules==0.2.4
rsa==4.0
six==1.12.0
